+++ types/src/model/wallet.rs
@@ -424,7 +424,7 @@ pub struct ListReceivedByAddressItem {
     /// The total amount received by the address.
     pub amount: Amount,
     /// The number of confirmations of the most recent transaction included.
-    pub confirmations: i64, // Docs do not indicate what negative value means?
+    pub confirmations: u32, // Docs do not indicate what negative value means?
     /// The label of the receiving address. The default label is "".
     pub label: String,
     /// The ids of transactions received with the address.
@@ -492,7 +492,7 @@ pub struct ListSinceBlockTransaction {
     ///
     /// Available for 'send' and 'receive' category of transactions. When it's < 0, it means the
     /// transaction conflicted that many blocks ago.
-    pub confirmations: i64,
+    pub confirmations: u32,
     /// The block hash containing the transaction.
     ///
     /// Available for 'send' and 'receive' category of transactions.
@@ -556,7 +556,7 @@ pub struct ListTransactionsItem {
     /// The number of confirmations for the transaction.
     ///
     /// Negative confirmations indicate the transaction conflicts with the block chain.
-    pub confirmations: i64,
+    pub confirmations: u32,
     /// Whether we consider the outputs of this unconfirmed transaction safe to spend.
     pub trusted: bool,
     /// The block hash containing the transaction.
